Centroid distance calculation
So the result of this is that the centroid approach is only usable for the top or bottom of the paragraph.

The old approach worked well for the sides, but a new, left weighted centered height appraoch is going to have to be implemented for the left side as the centroid of words is often futher than top or bottom neighbors.